requirements.txt
fastapi>=0.110
uvicorn>=0.29
openai>=1.0
numpy>=1.24
python-dotenv>=1.0
requests>=2.31
